Temporary resource hacks, until I add proper resource handling.

+++ src/compositor/surface.c
@@ -271,10 +271,8 @@ wlc_surface_free(struct wlc_surface *surface)
    if (surface->resource)
       wl_resource_destroy(surface->resource);
 
-   if (surface->compositor) {
-      if (surface->compositor->render)
-         surface->compositor->render->api.destroy(surface);
-   }
+   if (surface->compositor && surface->compositor->render)
+      surface->compositor->render->api.destroy(surface);
 
    free(surface);
 }
